General clean up
